requirements.txt
absl-py==0.8.1
amply==0.1.4
art==4.5
astroid==2.4.2
certifi==2020.6.20
cffi==1.13.0
click==8.0.1
coverage==5.3
cycler==0.10.0
decorator==4.4.1
docutils==0.17.1
grpcio==1.16.1
imageio==2.6.1
importlib-metadata==4.8.1
imutils==0.5.4
isort==4.3.21
joblib==0.16.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
matplotlib==3.3.2
mccabe==0.6.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.4
numpy==1.17.2
olefile==0.46
opencv-contrib-python-headless==4.5.3.56
Pillow==6.2.0
progress==1.5
protobuf==3.10.1
PuLP==2.4
pycm==3.0
pycparser==2.19
pylint==2.5.3
pyparsing==2.4.2
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
python-dateutil==2.8.0
python-dotenv==0.19.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
schedule==1.1.0
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.12.0
tensorboard==2.0.0
threadpoolctl==2.1.0
toml==0.10.1
torch==1.3.0
torchsummary==1.5.1
torchvision==0.4.1a0+d94043a
tornado==6.0.3
typed-ast==1.4.1
typing-extensions==3.10.0.2
Werkzeug==0.16.0
wrapt==1.12.1
yacs==0.1.8
zipp==3.6.0
